pop: move struct email to new types header file
The type will be shared with other code so moving to a .h file is helpful.
